pyproject.toml
[project]
name = "ocrtransforms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Joe Barrow", email = "joseph.d.barrow@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "formalpdf>=0.1.6",
    "pytest>=8.4.2",
]
